+++ packages/instantsearch.js/src/__tests__/common-widgets.test.tsx
@@ -37,7 +37,7 @@ import {
   chat,
   chatTrigger,
   promptSuggestions,
-  EXPERIMENTAL_autocomplete,
+  autocomplete,
   filterSuggestions,
 } from '../widgets';
 
@@ -717,10 +717,10 @@ const testSetups: TestSetupsMap<TestSuites, 'javascript'> = {
   createAutocompleteWidgetTests({ instantSearchOptions, widgetParams }) {
     instantsearch(instantSearchOptions)
       .addWidgets([
-        EXPERIMENTAL_autocomplete({
+        autocomplete({
           container: document.body.appendChild(document.createElement('div')),
           ...widgetParams,
-        } as Parameters<typeof EXPERIMENTAL_autocomplete>[0]),
+        } as Parameters<typeof autocomplete>[0]),
       ])
       .on('error', () => {
         /*