+++ packages/react-instantsearch/src/widgets/__tests__/Autocomplete.test.tsx
@@ -13,11 +13,11 @@ import { act, fireEvent, render } from '@testing-library/react';
 import React from 'react';
 import { InstantSearchSSRProvider } from 'react-instantsearch-core';
 
-import { EXPERIMENTAL_Autocomplete } from '../Autocomplete';
+import { Autocomplete } from '../Autocomplete';
 
 const noop = () => {};
 
-describe('EXPERIMENTAL_Autocomplete', () => {
+describe('Autocomplete', () => {
   beforeEach(() => {
     Object.defineProperty(window, 'matchMedia', {
       writable: true,
@@ -49,7 +49,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
             compositionID="my-comp"
             indexName={undefined as unknown as string}
           >
-            <EXPERIMENTAL_Autocomplete
+            <Autocomplete
               feeds={[
                 {
                   feedID: 'products',
@@ -76,7 +76,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
       expect(() => {
         render(
           <InstantSearchTestWrapper searchClient={searchClient}>
-            <EXPERIMENTAL_Autocomplete
+            <Autocomplete
               feeds={[
                 {
                   feedID: 'products',
@@ -102,7 +102,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
       expect(() => {
         render(
           <InstantSearchTestWrapper searchClient={searchClient}>
-            <EXPERIMENTAL_Autocomplete
+            <Autocomplete
               indices={[
                 {
                   indexName: 'my-index',
@@ -136,7 +136,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
             searchClient={searchClient}
             indexName="indexName"
           >
-            <EXPERIMENTAL_Autocomplete
+            <Autocomplete
               indices={indices}
               requiresSearch={requiresSearch}
             />
@@ -171,7 +171,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
           searchClient={searchClient}
           indexName="indexName"
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             indices={[
               {
                 indexName: 'my-index',
@@ -203,7 +203,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
           searchClient={searchClient}
           indexName="indexName"
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             indices={[
               {
                 indexName: 'my-index',
@@ -244,7 +244,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
             indexName: { query: 'macbook' },
           }}
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             indices={[
               {
                 indexName: 'my-index',
@@ -293,7 +293,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
             indexName: { query: 'macbook' },
           }}
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             indices={[
               {
                 indexName: 'my-index',
@@ -344,7 +344,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
             searchClient={searchClient as never}
             indexName="instant_search"
           >
-            <EXPERIMENTAL_Autocomplete
+            <Autocomplete
               indices={[
                 {
                   indexName: 'instant_search',
@@ -387,7 +387,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
           searchClient={searchClient}
           indexName="indexName"
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             autoFocus
             indices={[
               {
@@ -420,7 +420,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
           searchClient={searchClient}
           indexName="indexName"
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             indices={[
               {
                 indexName: 'my-index',
@@ -471,7 +471,7 @@ describe('EXPERIMENTAL_Autocomplete', () => {
           searchClient={searchClient}
           indexName="indexName"
         >
-          <EXPERIMENTAL_Autocomplete
+          <Autocomplete
             detachedMediaQuery={detachedMediaQuery}
             indices={[
               {