+++ packages/instantsearch.js/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.108.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.107.0...instantsearch.js@4.108.0) (2026-07-23)
+
+
+### Bug Fixes
+
+* **chat:** allow tool result submission during response streaming ([#7118](https://github.com/algolia/instantsearch/issues/7118)) ([2d196b3](https://github.com/algolia/instantsearch/commit/2d196b3c003c616e2a0a0fd57b800cef7125f641))
+
+
+### Features
+
+* **autocomplete:** autocomplete widget and isolated index are stable ([#7117](https://github.com/algolia/instantsearch/issues/7117)) ([8fd77dc](https://github.com/algolia/instantsearch/commit/8fd77dce8dbb9a641facfe1639cb678f968cdd46))
+
+
+
+
+
+# [4.107.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.106.0...instantsearch.js@4.107.0) (2026-07-21)
+
+
+### Bug Fixes
+
+* **chat:** allow user to override `sendAutomaticallyWhen` in the Chat widget ([#7116](https://github.com/algolia/instantsearch/issues/7116)) ([1c82852](https://github.com/algolia/instantsearch/commit/1c82852ae4ff874f35c4ac82724d59952b71cd7d))
+* **index:** clear parent only on removed widgets ([#7115](https://github.com/algolia/instantsearch/issues/7115)) ([c8e392c](https://github.com/algolia/instantsearch/commit/c8e392c23faa45e810757338380770811692c16b))
+
+
+### Features
+
+* allow chat and autocomplete to opt out of main search ([#7111](https://github.com/algolia/instantsearch/issues/7111)) ([997a511](https://github.com/algolia/instantsearch/commit/997a511f7ca034c92fcc53bd707c6448cf1bfdcf))
+
+
+
+
+
 # [4.106.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.105.0...instantsearch.js@4.106.0) (2026-07-14)